Moving code out of __init__.py files
