+++ src/spellbot/interactions/ban_interaction.py
@@ -4,8 +4,9 @@ from typing import Optional
 from discord.ext import commands
 
 from spellbot.client import SpellBot
-from spellbot.interactions import BaseInteraction
 from spellbot.operations import safe_send_user
+
+from .base_interaction import BaseInteraction
 
 logger = logging.getLogger(__name__)
