+++ src/spellbot/interactions/config_interaction.py
@@ -8,12 +8,13 @@ from discord_slash.context import ComponentContext, InteractionContext
 from discord_slash.model import ButtonStyle
 from pygicord import Config, Paginator
 
-from spellbot.client import SpellBot
-from spellbot.interactions import BaseInteraction
+from spellbot import SpellBot
 from spellbot.models import Channel
 from spellbot.operations import safe_send_channel, safe_update_embed_origin
 from spellbot.settings import Settings
 from spellbot.utils import EMBED_DESCRIPTION_SIZE_LIMIT, log_warning
+
+from .base_interaction import BaseInteraction
 
 logger = logging.getLogger(__name__)
