+++ src/spellbot/interactions/score_interaction.py
@@ -4,10 +4,11 @@ from typing import Union
 import discord
 from discord_slash.context import InteractionContext
 
-from spellbot.client import SpellBot
-from spellbot.interactions import BaseInteraction
+from spellbot import SpellBot
 from spellbot.operations import safe_send_channel
 from spellbot.settings import Settings
+
+from .base_interaction import BaseInteraction
 
 logger = logging.getLogger(__name__)
