+++ tests/cogs/test_score.py
@@ -4,8 +4,9 @@ import discord
 import pytest
 from discord_slash.context import InteractionContext, MenuContext
 
-from spellbot import Settings, SpellBot
+from spellbot import SpellBot
 from spellbot.cogs.score import ScoreCog
+from spellbot.settings import Settings
 from tests.factories.game import GameFactory
 from tests.factories.play import PlayFactory
 from tests.mocks import build_channel, build_ctx, build_guild
