+++ tests/fixtures.py
@@ -10,7 +10,7 @@ import pytest
 from aiohttp.client import ClientSession
 from discord_slash.context import InteractionContext
 
-from spellbot.client import SpellBot, build_bot
+from spellbot import SpellBot, build_bot
 from spellbot.database import (
     DatabaseSession,
     db_session_maker,
