+++ tests/test_spelltable.py
@@ -7,7 +7,8 @@ import pytest
 from aiohttp.client_exceptions import ClientError
 from aiohttp_retry.client import RetryClient
 
-from spellbot import Settings, spelltable
+from spellbot import spelltable
+from spellbot.settings import Settings
 from spellbot.spelltable import generate_link
